typecast-python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "typecast-python"
version = "0.1.5"
description = "Official Typecast Python SDK - Convert text to lifelike speech using AI-powered voices"
authors = [
    {name = "Neosapience", email = "help@typecast.ai"}
]
maintainers = [
    {name = "Neosapience", email = "help@typecast.ai"}
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.11"
keywords = ["typecast", "tts", "text-to-speech", "ai", "voice", "audio", "speech-synthesis", "voice-generation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Operating System :: OS Independent",
]
dependencies = [
    "aiohttp>=3.8.0",
    "requests>=2.28.0",
    "pydantic>=2.0.0",
    "sseclient-py>=1.7.2",
    "websockets>=10.0",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "isort>=5.0.0",
    "pytest-mock>=3.14.0",
    "pytest-asyncio>=0.21.0",
]

[project.urls]
Homepage = "https://typecast.ai"
Documentation = "https://typecast.ai/docs/overview"
Repository = "https://github.com/neosapience/typecast-python"
Issues = "https://github.com/neosapience/typecast-python/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/typecast"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/LICENSE",
    "/README.md",
    "/pyproject.toml",
]
exclude = [
    "/.env*",
    "/tests",
    "/examples",
    "/spec",
    "/docker",
    "*.lock",
    "/Makefile",
    "/MANIFEST.in",
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.14.0",
]
